BUILD : modify to getting argument from cmdline with make
to get the value of some variable calculated outside of a C program,
it is need to get the value from the command line.

in this case you need DEFINED_FILE_COUNT which means the number of
hashes of files. that decide the size of myhashes sector, which
statically stores the hash of the file at compile time.