LoopOptTutorial: Fix misspecified basicblock
Summarize what was done in the previous commit,
1. Copy the loop after the loop's Preheader.
2. Update the values of the copied loop.
3. Change the Preheader of the loop to which you copied the Preheader
   values in the Predecessor of the Preheader.

Because we used SplitBlock, there is now a Splitted Preheader between
the loop and the Preheader. Therefore, you must change the Splitted
Preheader that exists in the Preheader to the Preheader of the copied
loop, rather than the Preheader value that exists in the Predecessor
of the Preheader.