LoopOptTutorial: Remove code for updating domtree
previous summary of codes.
1. splitted loop.
2. insert cloned loop into splitted block.

we called `myCloneLoopWithPreheader` for cloning and insert into.
in `myCloneLoopWithPreheader`, not only cloning and insert but also
update dominator tree. without updating dominator tree, this works
will be failed.

but every calling `myCloneLoopWithPreheader` will update dominator
tree inside function. and it is not necessary obviously and waste
time as much as you calling it.

therefore, we need to learn how to deal with dominator when we modify
or manipulator from it. this will be the main focus of future commits.